pyproject.toml
[tool.poetry]
name = "svc-infra"
version = "0.1.709"
description = "Infrastructure for building and deploying prod-ready services"
authors = ["Ali Khatami <aliikhatami94@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "svc_infra", from = "src" }]
keywords = ["fastapi", "sqlalchemy", "alembic", "auth", "infra", "async", "pydantic"]
include = [
    "src/svc_infra/py.typed",
    "src/svc_infra/db/templates/**/*",
    "src/svc_infra/obs/templates/**/*",
    "src/svc_infra/obs/grafana/templates/**/*",
    "src/svc_infra/obs/grafana/dashboards/**/*",
    "src/svc_infra/docs/**/*"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
    "Typing :: Typed"
]

[tool.poetry.urls]
Homepage = "https://github.com/nfraxlab/svc-infra"
Repository = "https://github.com/nfraxlab/svc-infra"
Issues = "https://github.com/nfraxlab/svc-infra/issues"
Documentation = "https://nfrax.com/svc-infra"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# Core
fastapi = ">=0.110.0"
pydantic-settings = ">=2.0"
python-dotenv = ">=1.0.0"
typer = ">=0.12.0"
httpx = ">=0.25.0"

# DB + auth
sqlalchemy = { extras = ["asyncio"], version = ">=2.0" }
alembic = ">=1.13.0"
greenlet = ">=3.0"
fastapi-users = { extras = ["oauth"], version = ">=14.0.0" }
fastapi-users-db-sqlalchemy = ">=7.0.0"
authlib = ">=1.0.0"
httpx-oauth = ">=0.15.0"
passlib = { extras = ["bcrypt"], version = ">=1.7.4" }
email-validator = ">=2.0.0"

# Database drivers (all optional)
psycopg = {extras = ["binary"], version = ">=3.0"}   # v3
asyncpg = { version = ">=0.29.0", optional = true }
aiosqlite = { version = ">=0.19.0", optional = true }
pymysql = { version = ">=1.1.0", optional = true }
mysqlclient = { version = ">=2.2.0", optional = true }
pyodbc = { version = ">=5.0.0", optional = true }
snowflake-connector-python = { version = ">=3.0.0", optional = true }
redshift-connector = { version = ">=2.0.0", optional = true }
duckdb = { version = ">=0.10.0", optional = true }

# Storage
aiofiles = ">=24.0.0"
aioboto3 = { version = ">=12.0.0", optional = true }

# Other infra deps
mcp = ">=1.0.0"
opentelemetry-sdk = ">=1.20.0"
opentelemetry-instrumentation-fastapi = ">=0.41b0"
opentelemetry-instrumentation-sqlalchemy = ">=0.41b0"
opentelemetry-instrumentation-requests = ">=0.41b0"
opentelemetry-instrumentation-httpx = ">=0.41b0"
opentelemetry-exporter-otlp = ">=1.20.0"
opentelemetry-propagator-b3 = ">=1.20.0"
prometheus-client = ">=0.18.0"
redis = ">=5.0.0"
cashews = {extras = ["redis"], version = ">=7.0"}
itsdangerous = ">=2.0.0"
pyotp = ">=2.9.0"
websockets = ">=12.0"

# Optional dependencies (install via extras)
motor = { version = ">=3.0.0", optional = true }
stripe = { version = ">=7.0.0", optional = true }
adyen = { version = ">=10.0.0", optional = true }
psycopg2-binary = { version = ">=2.9.0", optional = true }

[tool.poetry.extras]
# Postgres v3 (recommended)
pg = ["psycopg", "asyncpg"]
# Postgres v2 (legacy compatibility)
pg2 = ["psycopg2-binary"]
sqlite = ["aiosqlite"]
mysql = ["pymysql", "mysqlclient"]
# Payment providers
payments = ["stripe", "adyen"]
stripe = ["stripe"]
adyen = ["adyen"]
# MongoDB
mongodb = ["motor"]
mssql = ["pyodbc"]
snowflake = ["snowflake-connector-python"]
redshift = ["redshift-connector"]
duckdb = ["duckdb"]
metrics = ["prometheus-client"]
s3 = ["aioboto3"]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
ruff = ">=0.6.0"
mypy = ">=1.10.0"
types-requests = ">=2.31.0"
pytest-mock = ">=3.12.0"
httpx = ">=0.25.0"
fakeredis = ">=2.20.0"
moto = {extras = ["s3"], version = ">=5.0.0"}
aiosqlite = ">=0.19.0"  # Required for acceptance tests using SQLite
pre-commit = ">=3.0.0"
# Documentation
mkdocs = ">=1.6.0"
mkdocs-material = ">=9.5.0"
mkdocstrings = {extras = ["python"], version = ">=0.25.0"}

[tool.poetry.scripts]
svc-infra = "svc_infra.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*",]
python_functions = ["test_*"]
markers = [
    "acceptance: End-to-end acceptance tests running against the acceptance app or BASE_URL",
    "unit: Unit tests",
    "security: Security and auth hardening tests",
    "ratelimit: Rate limiting and abuse protection tests",
    "concurrency: Idempotency and concurrency control tests",
    "jobs: Background jobs and scheduling tests",
    "webhooks: Webhooks framework tests",
    "billing: Billing primitives tests",
    "tenancy: Tenancy isolation and enforcement tests",
    "data_lifecycle: Data lifecycle (fixtures, retention, erasure, backups)",
    "ops: SLOs & Ops tests (probes, breaker, instrumentation)",
    "dx: Developer experience and quality gates tests",
    "admin: Admin scope and impersonation tests",
    "storage: File storage system tests",
]
filterwarnings = [
    "ignore:The `route` decorator is deprecated:DeprecationWarning:starlette.*",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",    # line too long (handled by formatter)
    "RUF022",  # __all__ not sorted (we use logical grouping)
    "RUF012",  # mutable class attrs (handled by dataclass/pydantic)
    "B007",    # unused loop variable (often intentional)
    "B008",    # function call in default arg (typer.Option pattern)
    "B027",    # empty method in ABC (intentional for optional callbacks)
    "B904",    # raise from err (not always clearer)
    "B905",    # zip strict (Python 3.10+ only, adds verbosity)
    "SIM102",  # nested if statements (sometimes clearer)
    "SIM103",  # return condition directly (sometimes clearer with if)
    "SIM105",  # contextlib.suppress (explicit try/except often clearer)
    "SIM108",  # ternary operator (readability preference)
    "SIM110",  # use all() (sometimes clearer with explicit loop)
    "SIM113",  # use enumerate (sometimes clearer without)
    "SIM115",  # context manager for files (not always applicable)
    "SIM117",  # single with statement (sometimes clearer separate)
    "SIM118",  # key in dict vs dict.keys() (minor)
    "UP028",   # yield from (sometimes clearer with explicit yield)
    "UP038",   # isinstance(x, (A, B)) - deprecated rule, requires unsafe fixes
    "RUF001",  # ambiguous unicode char in string (intentional in docs)
    "RUF002",  # ambiguous unicode char in docstring (intentional curly quotes)
    "RUF003",  # ambiguous unicode char in comment (intentional curly quotes)
    "B023",    # function uses loop variable (often intentional in closures)
    "B039",    # contextvar default mutable (handled by contextvar)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "B018",    # useless expression in tests is fine
]

[tool.ruff.lint.isort]
known-first-party = ["svc_infra"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
